Module dusk_plonk::circuit

Tools & traits for PLONK circuits

Structs

PublicInputValue

Structure that represents a PLONK Circuit Public Input converted into it’s &[BlsScalar] repr.

VerifierData

Collection of structs/objects that the Verifier will use in order to de/serialize data needed for Circuit proof verification. This structure can be seen as a link between the Circuit public input positions and the VerifierKey that the Verifier needs to use.

Traits

Circuit

Trait that should be implemented for any circuit function to provide to it the capabilities of automatically being able to generate, and verify proofs as well as compile the circuit.