Module dusk_plonk::commitment_scheme::kzg10

Implementation of the KZG10 polynomial commitment scheme.

Re-exports

pub use key::CommitKey;

pub use key::OpeningKey;

pub use srs::PublicParameters;

Modules

errors
Errors related to KZG10

key
Key module contains the utilities and data structures that support the generation and usage of Commit and Opening keys.

srs
The Public Parameters can also be referred to as the Structured Reference String (SRS).

Structs

AggregateProof
Proof that multiple polynomials were correctly evaluated at a point z, each producing their respective evaluated points p_i(z).

Commitment
Holds a commitment to a polynomial in a form of a G1Affine Bls12_381 point.

Proof
Proof that a polynomial p was correctly evaluated at a point z producing the evaluated point p(z).