Struct dusk_plonk::commitment_scheme::kzg10::key::OpeningKey

pub struct OpeningKey {
    pub g: G1Affine,
    pub h: G2Affine,
    pub beta_h: G2Affine,
    pub prepared_h: G2Prepared,
    pub prepared_beta_h: G2Prepared,
}

Opening Key is used to verify opening proofs made about a committed polynomial.

Fields
g: G1Affine

The generator of G1.

h: G2Affine

The generator of G2.

beta_h: G2Affine

\beta times the above generator of G2.

prepared_h: G2Prepared

The generator of G2, prepared for use in pairings.

prepared_beta_h: G2Prepared

\beta times the above generator of G2, prepared for use in pairings.

Implementations

impl OpeningKey

pub fn check(&self, point: Scalar, proof: Proof) -> bool

Checks that a polynomial p was evaluated at a point z and returned the value specified v. ie. v = p(z).

pub fn batch_check(
&self,
points: &[Scalar],
proofs: &[Proof],
transcript: &mut Transcript
) -> Result<(), Error>

Checks whether a batch of polynomials evaluated at different points, returned their specified value.