Module dusk_plonk::commitment_scheme::kzg10::srs

The Public Parameters can also be referred to as the Structured Reference String (SRS).

Structs

PublicParameters
The Public Parameters can also be referred to as the Structured Reference String (SRS). It is available to both the prover and verifier and allows the verifier to efficiently verify and make claims about polynomials up to and including a configured degree.