Struct dusk_plonk::commitment_scheme::kzg10::AggregateProof

pub struct AggregateProof {
    pub commitment_to_witness: Commitment,
    pub evaluated_points: Vec<Scalar>,
    pub commitments_to_polynomials: Vec<Commitment>,
}

Proof that multiple polynomials were correctly evaluated at a point z, each producing their respective evaluated points p_i(z).

Fields
commitment_to_witness: Commitment

This is a commitment to the aggregated witness polynomial.

evaluated_points: Vec<Scalar>

These are the results of the evaluating each polynomial at the point z.

commitments_to_polynomials: Vec<Commitment>

These are the commitments to the polynomials which you want to prove a statement about.

Implementations

impl AggregateProof

pub fn with_witness(witness: Commitment) -> AggregateProof

Initialises an AggregatedProof with the commitment to the witness.

pub fn add_part(&mut self, part: (Scalar, Commitment))

Adds an evaluated point with the commitment to the polynomial which produced it.

pub fn flatten(&self, transcript: &mut Transcript) -> Proof

Flattens an AggregateProof into a Proof. The transcript must have the same view as the transcript that was used to aggregate the witness in the proving stage.