Module dusk_plonk::constraint_system

The constraint System module stores the implementation of the PLONK Standard Composer, as well as the circuit tools and abstractions, used by the Composer to generate, build, preprocess circuits.

Modules

arithmetic
Simple Arithmetic gates

boolean
Boolean gate

ecc
Scalar multiplication gate

logic
XOR and AND gates

range
Range gate

Structs

StandardComposer
A composer is a circuit builder and will dictate how a circuit is built We will have a default Composer called StandardComposer

Variable
The value is a reference to the actual value that was added to the constraint system

Enums

WireData
Stores the data for a specific wire in an arithmetic circuit This data is the gate index and the type of wire Left(1) signifies that this wire belongs to the first gate and is the left wire