Module dusk_plonk::notes

This module is a self contained file which explains how PLONK and its protocol components work in our library.

The following section explains the workings of PLONK components:

Commitment schemes

Proof construction components

Permutation polynomials

This section explains how elements of the aggregated PLONK protocol work and some of the unique features in this library:

Prover and Verify algorithms

Unbalanced permutation arguments

KZG10 commitment scheme

Modules

commitment_schemes
This module explains the inner workings of commitment schemes.

kzg10_docs
In this module we show how, and why, the KZG10 polynomial commitment scheme has been implemented for this PLONK implementation.

permutation_arguments
Contained within this module are the notes on how the permutation argument is formed.

prove_verify
This module contains the notes on how the prover algorithm is constructed for PLONK.

snark_construction
This module contains the methodology of how zk-SNARKS are constructed.

unbalanced_perm_args
Contained within this module are the notes on how the permutation argument is formed.