Module dusk_plonk::plonkup

Module containing the plonkup works. Plonkup is the protcol for using precomputed and stored tables of values for specific functions to determine the output of gates within a circuit, without computing them.

Modules

constants

Module containing constants required in reinforced concrete

Structs

LookupTable

This struct is a table, contaning a vector, of arity 4 where each of the values is a BlsScalar. The elements of the table are determined by the function g for g(x,y), used to compute tuples.

MultiSet

MultiSet is struct containing vectors of scalars, which individually represents either a wire value or an index of a Plonkup table

PreprocessedLookupTable

This table will be the preprocessed version of the precomputed table, T, with arity 4. This structure is passed to the proof alongside the table of witness values.

WitnessTable

This witness table contains quieries to a lookup table for lookup gates This table is of arity 3.