Module dusk_plonk::prelude

Collection of functions needed to use plonk library.

Use this as the only import that you need to interact with the principal data structures of the plonk library.

Re-exports

pub use crate::commitment_scheme::kzg10::key::CommitKey;

pub use crate::commitment_scheme::kzg10::key::OpeningKey;

pub use crate::commitment_scheme::kzg10::PublicParameters;

pub use crate::proof_system::Proof;

pub use crate::proof_system::Prover;

pub use crate::proof_system::Verifier;

Modules

plonk_errors
Collection of errors that the library exposes/uses.

Structs

BlsScalar
Re-exported dusk-bls12_381::Scalar. Represents an element of the scalar field $\mathbb{F}_q$ of the BLS12-381 elliptic curve construction.

JubJubScalar
Re-exported dusk-jubjub::Scalar. Represents an element of the scalar field $\mathbb{F}_r$ of the Jubjub elliptic curve construction.

ProverKey
PLONK circuit proving key

StandardComposer
A composer is a circuit builder and will dictate how a circuit is built We will have a default Composer called StandardComposer

Variable
The value is a reference to the actual value that was added to the constraint system

VerifierKey
PLONK circuit verification key