Module dusk_plonk::prelude::plonk_errors

Collection of errors that the library exposes/uses.

Re-exports

pub use crate::commitment_scheme::kzg10::errors::KZG10Errors;

pub use crate::commitment_scheme::kzg10::errors::PolyCommitSchemeError;

Structs

FFTError
Represents an error triggered on any of the FFT module operations such as Polynomial or EvaluationDomain

ProofError
Represents an error triggered on any of the proof_system module operations such as verification errors

ProvingError
Represents an error on the Proving stage.

Enums

FFTErrors
Defines all of the possible FFTError types that we could have when we are working with the fft module.

PreProcessingError
Represents an error on the Circuit preprocessing stage.

ProofErrors
Defines all of the possible ProofError types that we could have when we are working with the proof_system module.