Struct dusk_plonk::prelude::BlsScalar

pub struct BlsScalar(pub [u64; 4]);

Represents an element of the scalar field $\mathbb{F}_q$ of the BLS12-381 elliptic curve construction.

Tuple Fields
0: [u64; 4]
Implementations

impl Scalar

pub const fn zero() -> Scalar

Returns zero, the additive identity.

pub const fn one() -> Scalar

Returns one, the multiplicative identity.

pub fn is_zero(&self) -> Choice

Checks in ct_time whether a Scalar is equal to zero.

pub fn is_one(&self) -> Choice

Checks in ct_time whether a Scalar is equal to one.

pub const fn internal_repr(&self) -> &[u64; 4]

Returns the internal representation of the Scalar.

pub const fn double(&self) -> Scalar

Doubles this field element.

pub fn to_bits(&self) -> [u8; 256]

Returns the bit representation of the given Scalar as an array of 256 bits represented as u8.

pub fn from_bytes_wide(bytes: &[u8; 64]) -> Scalar

Converts a 512-bit little endian integer into a Scalar by reducing by the modulus.

pub const fn from_raw(val: [u64; 4]) -> Scalar

Converts from an integer represented in little endian into its (congruent) Scalar representation.

pub fn random<T>(rand: &mut T) -> Scalar where
T: RngCore + CryptoRng,

Generate a valid Scalar choosen uniformly using user- provided rng.

By rng we mean any Rng that implements: Rng + CryptoRng.

pub fn reduce(&self) -> Scalar

Reduces the scalar and returns it multiplied by the montgomery radix.

pub const fn square(&self) -> Scalar

Squares this element.

pub fn sqrt(&self) -> CtOption<Scalar>

Computes the square root of this element, if it exists.

pub fn pow(&self, by: &[u64; 4]) -> Scalar

Exponentiates self by by, where by is a little-endian order integer exponent.

pub fn pow_vartime(&self, by: &[u64; 4]) -> Scalar

Exponentiates self by by, where by is a little-endian order integer exponent.

This operation is variable time with respect to the exponent. If the exponent is fixed, this operation is effectively constant time.

pub fn pow_of_2(by: u64) -> Scalar

Computes 2^X where X is a u64 without the need to generate

pub fn invert(&self) -> CtOption<Scalar>

Computes the multiplicative inverse of this element, failing if the element is zero.

pub const fn mul(&self, rhs: &Scalar) -> Scalar

Multiplies rhs by self, returning the result.

pub const fn sub(&self, rhs: &Scalar) -> Scalar

Subtracts rhs from self, returning the result.

pub const fn add(&self, rhs: &Scalar) -> Scalar

Adds rhs to self, returning the result.

pub const fn neg(&self) -> Scalar

Negates self.

pub fn divn(&mut self, n: u32)

SHR impl