Struct dusk_plonk::prelude::PublicParameters

pub struct PublicParameters { /* fields omitted */ }

The Public Parameters can also be referred to as the Structured Reference String (SRS). It is available to both the prover and verifier and allows the verifier to efficiently verify and make claims about polynomials up to and including a configured degree.

Implementations

impl PublicParameters

pub fn commit_key(&self) -> &CommitKey

Returns an untrimmed CommitKey reference contained in the PublicParameters instance.

pub fn opening_key(&self) -> &OpeningKey

Returns an OpeningKey reference contained in the PublicParameters instance.

pub fn setup<R: RngCore + CryptoRng>(
max_degree: usize,
rng: &mut R
) -> Result<PublicParameters, Error>

Setup generates the public parameters using a random number generator. This method will in most cases be used for testing and exploration. In reality, a Trusted party or a Multiparty Computation will be used to generate the SRS. Returns an error if the configured degree is less than one.

pub fn to_raw_var_bytes(&self) -> Vec<u8>

Serialize the PublicParameters into bytes.

This operation is designed to store the raw representation of the contents of the PublicParameters. Therefore, the size of the bytes outputed by this function is expected to be the double than the one that PublicParameters::to_var_bytes.

Note

This function should be used when we want to serialize the PublicParameters allowing a really fast deserialization later. This functions output should not be used by the regular PublicParameters::from_slice fn.

pub unsafe fn from_slice_unchecked(bytes: &[u8]) -> Self

Deserialize PublicParameters from a set of bytes created by PublicParameters::to_raw_var_bytes.

The bytes source is expected to be trusted and no checks will be performed reggarding the content of the points that the bytes contain serialized.

Safety

This function will not produce any memory errors but can deal to the generation of invalid or unsafe points/keys. To make sure this does not happen, the inputed bytes must match the ones that were generated by the encoding functions of this lib.

pub fn to_var_bytes(&self) -> Vec<u8>

Serialises a PublicParameters struct into a slice of bytes.

pub fn from_slice(bytes: &[u8]) -> Result<PublicParameters, Error>

Deserialise a slice of bytes into a Public Parameter struct performing security and consistency checks for each point that the bytes contain.

Note

This function can be really slow if the PublicParameters have a certain degree. If the bytes come from a trusted source such as a local file, we recommend to use PublicParameters::from_slice_unchecked and PublicParameters::to_raw_var_bytes.

pub fn trim(
&self,
truncated_degree: usize
) -> Result<(CommitKey, OpeningKey), Error>

Trim truncates the PublicParameters to allow the prover to commit to polynomials up to the and including the truncated degree. Returns the CommitKey and OpeningKey used to generate and verify proofs.

Returns an error if the truncated degree is larger than the public parameters configured degree.

pub fn max_degree(&self) -> usize

Max degree specifies the largest Polynomial that this prover key can commit to.