Module dusk_plonk::proof_system

proving system

Re-exports

pub use proof::Proof;

pub use prover::Prover;

pub use verifier::Verifier;

Modules

proof
Represents a PLONK Proof A Proof stores the commitments to all of the elements that are needed to univocally identify a prove of some statement.

prover
Represents a PLONK Prover

verifier
Represents a PLONK Verifier